🌈 Changed size types, cleanup

+++ Nova/IWebApplication.h
@@ -47,7 +47,7 @@ namespace Neko
             {
             }
             
-            inline int32 ProcessRequest(Http::RequestData& request, Http::ResponseData& response)
+            inline int16 ProcessRequest(Http::RequestData& request, Http::ResponseData& response)
             {
                 return Context.Execute(request, response);
             }
