+++ SocketSSL.h
@@ -56,9 +56,9 @@ namespace Neko
             /** Initiates from existing SSL connection. */
 			SocketSSL(const Net::INetSocket& socket, SSL* connection);
 			
-            virtual long GetPacketBlocking(void* buffer, const ulong length, const uint32& timeout) const override;
+            virtual long GetPacketBlocking(void* buffer, const ulong length, const int32& timeout) const override;
             
-            virtual long SendAllPacketsWait(const void* buffer, const ulong length, const uint32& timeout) const override;
+            virtual long SendAllPacketsWait(const void* buffer, const ulong length, const int32& timeout) const override;
             
             virtual void Close() override;
             
@@ -68,7 +68,7 @@ namespace Neko
             virtual NEKO_FORCE_INLINE void* GetTlsSession() const override { return (void* )Connection; };
             
             /** Higher level SSL connect */
-            int16 Connect();
+            int32 Connect();
             
 			bool Handshake();
